# backend/openapi.json
{
  "components": {
    "schemas": {
      "CaregiverAcceptRequest": {
        "additionalProperties": false,
        "properties": {
          "invite_token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Invite Token",
            "type": "string"
          }
        },
        "required": [
          "invite_token"
        ],
        "title": "CaregiverAcceptRequest",
        "type": "object"
      },
...
      "CaregiverLinkResponse": {
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "caregiver_email": {
            "format": "email",
            "title": "Caregiver Email",
            "type": "string"
          },
          "caregiver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caregiver Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "permission_level": {
            "$ref": "#/components/schemas/CaregiverPermission"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "status": {
            "$ref": "#/components/schemas/CaregiverLinkStatus"
          }
        },
        "required": [
          "id",
          "patient_id",
          "caregiver_id",
          "caregiver_email",
          "relationship",
          "permission_level",
          "status",
          "created_at",
          "accepted_at",
          "revoked_at"
        ],
        "title": "CaregiverLinkResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "$defs": {
          "ErrorDetail": {
            "properties": {
              "code": {
                "examples": [
                  "validation_error"
                ],
                "title": "Code",
                "type": "string"
              },
              "details": {
                "default": null,
                "title": "Details"
              },
              "message": {
                "examples": [
                  "Request validation failed"
                ],
                "title": "Message",
                "type": "string"
              },
              "request_id": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "default": null,
                "examples": [
                  "01J00000000000000000000000"
                ],
                "title": "Request Id"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "title": "ErrorDetail",
            "type": "object"
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LinkedPatientResponse": {
        "properties": {
          "link": {
            "$ref": "#/components/schemas/CaregiverLinkResponse"
          },
          "patient": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "link",
          "patient"
        ],
        "title": "LinkedPatientResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "accessibility_preferences": {
            "additionalProperties": true,
            "title": "Accessibility Preferences",
            "type": "object"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "PATIENT"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "ScanLabelRequest": {
        "additionalProperties": false,
        "properties": {
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          }
        },
        "required": [
          "image_id"
        ],
        "title": "ScanLabelRequest",
        "type": "object"
      },
      "ScanLabelResponse": {
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "extracted_fields": {
            "additionalProperties": true,
            "title": "Extracted Fields",
            "type": "object"
          },
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "requires_confirmation": {
            "const": true,
            "default": true,
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "result": {
            "enum": [
              "LABEL_READ",
              "UNREADABLE_IMAGE"
            ],
            "title": "Result",
            "type": "string"
          },
          "safety_message": {
            "title": "Safety Message",
            "type": "string"
          }
        },
        "required": [
          "result",
          "ocr_text",
          "extracted_fields",
          "safety_message"
        ],
        "title": "ScanLabelResponse",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_expires_in": {
            "title": "Access Expires In",
            "type": "integer"
          },
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_expires_in": {
            "title": "Refresh Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "access_expires_in",
          "refresh_expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "accessibility_preferences": {
            "additionalProperties": true,
            "title": "Accessibility Preferences",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "timezone",
          "accessibility_preferences",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
      "VerifyMedicationRequest": {
        "additionalProperties": false,
        "properties": {
          "dose_log_id": {
            "format": "uuid",
            "title": "Dose Log Id",
            "type": "string"
          },
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "verification_type": {
            "enum": [
              "PILL_VERIFY",
              "BOTTLE_VERIFY"
            ],
            "title": "Verification Type",
            "type": "string"
          }
        },
        "required": [
          "dose_log_id",
          "image_id",
          "verification_type"
        ],
        "title": "VerifyMedicationRequest",
        "type": "object"
      },
      "VerifyMedicationResponse": {
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "requires_confirmation": {
            "const": true,
            "default": true,
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "result": {
            "$ref": "#/components/schemas/VerificationResult"
          },
          "safety_message": {
            "title": "Safety Message",
            "type": "string"
          }
        },
        "required": [
          "result",
          "safety_message"
        ],
        "title": "VerifyMedicationResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Medication assistance API. It does not provide medical advice.",
    "summary": "Production API contract for MedGuide AI.",
    "title": "MedGuide AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/ai/scan-label": {
      "post": {
        "operationId": "scan_medication_label_api_v1_ai_scan_label_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanLabelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scan Medication Label",
        "tags": [
          "AI assistance"
        ]
      }
    },
    "/api/v1/ai/verify-medication": {
      "post": {
        "operationId": "verify_medication_api_v1_ai_verify_medication_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMedicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyMedicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Medication",
        "tags": [
          "AI assistance"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/caregivers/accept": {
      "post": {
        "operationId": "accept_caregiver_invite_api_v1_caregivers_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaregiverAcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverLinkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Caregiver Invite",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/caregivers/invite": {
      "post": {
        "operationId": "invite_caregiver_api_v1_caregivers_invite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaregiverInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invite Caregiver",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/caregivers/links/{link_id}/revoke": {
      "patch": {
        "operationId": "revoke_caregiver_link_api_v1_caregivers_links__link_id__revoke_patch",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverLinkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Caregiver Link",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/caregivers/patients": {
      "get": {
        "operationId": "list_linked_patients_api_v1_caregivers_patients_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LinkedPatientResponse"
                  },
                  "title": "Response List Linked Patients Api V1 Caregivers Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Linked Patients",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/caregivers/patients/{patient_id}/dose-logs": {
      "get": {
        "operationId": "get_linked_patient_dose_logs_api_v1_caregivers_patients__patient_id__dose_logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoseLogResponse"
                  },
                  "title": "Response Get Linked Patient Dose Logs Api V1 Caregivers Patients  Patient Id  Dose Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Linked Patient Dose Logs",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/caregivers/patients/{patient_id}/today": {
      "get": {
        "operationId": "get_linked_patient_today_api_v1_caregivers_patients__patient_id__today_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientTodayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Linked Patient Today",
        "tags": [
          "caregivers"
        ]
      }
    },
    "/api/v1/devices": {
      "post": {
        "operationId": "register_device_api_v1_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "delete": {
        "operationId": "delete_device_api_v1_devices__device_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Device",
        "tags": [
          "devices"
        ]
      }
    },
...
  },
...
}